+++ Joker.Kafka/KSql/Query/Functions/Entry.cs
@@ -1,0 +1,8 @@
+﻿namespace Kafka.DotNet.ksqlDB.KSql.Query.Functions
+{
+  public struct Entry
+  {
+    public string K { get; set; }
+    public string V { get; set; }
+  }
+}